+++ android/src/main/java/vn/hunghd/flutterdownloader/DownloadWorker.java
@@ -12,6 +12,9 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
 import android.content.res.Resources;
+import android.database.Cursor;
+import android.media.MediaScannerConnection;
+import android.net.Uri;
 import android.os.Build;
 
 import androidx.annotation.NonNull;
@@ -25,6 +28,8 @@ import android.provider.MediaStore;
 import android.text.TextUtils;
 import android.util.Log;
 
+import androidx.annotation.RequiresApi;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -32,6 +37,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLDecoder;
@@ -65,6 +71,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
     public static final String ARG_OPEN_FILE_FROM_NOTIFICATION = "open_file_from_notification";
     public static final String ARG_CALLBACK_HANDLE = "callback_handle";
     public static final String ARG_DEBUG = "debug";
+    public static final String ARG_SAVE_IN_PUBLIC_STORAGE = "save_in_public_storage";
 
     private static final String TAG = DownloadWorker.class.getSimpleName();
     private static final int BUFFER_SIZE = 4096;
@@ -73,7 +80,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
 
     private static final AtomicBoolean isolateStarted = new AtomicBoolean(false);
     private static final ArrayDeque<List> isolateQueue = new ArrayDeque<>();
-    private static FlutterEngine flutterEngine;
+    private static FlutterEngine backgroundFlutterEngine;
 
     private final Pattern charsetPattern = Pattern.compile("(?i)\\bcharset=\\s*\"?([^\\s;\"]*)");
     private final Pattern filenameStarPattern = Pattern.compile("(?i)\\bfilename\\*=([^']+)'([^']*)'\"?([^\"]+)\"?");
@@ -89,6 +96,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
     private int primaryId;
     private String msgStarted, msgInProgress, msgCanceled, msgFailed, msgPaused, msgComplete;
     private long lastCallUpdateNotification = 0;
+    private boolean saveInPublicStorage;
 
     public DownloadWorker(@NonNull final Context context,
                           @NonNull WorkerParameters params) {
@@ -104,28 +112,29 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
 
     private void startBackgroundIsolate(Context context) {
         synchronized (isolateStarted) {
-            if (flutterEngine == null) {
+            if (backgroundFlutterEngine == null) {
                 SharedPreferences pref = context.getSharedPreferences(FlutterDownloaderPlugin.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE);
                 long callbackHandle = pref.getLong(FlutterDownloaderPlugin.CALLBACK_DISPATCHER_HANDLE_KEY, 0);
+
+                backgroundFlutterEngine = new FlutterEngine(getApplicationContext());
+
                 // We need to create an instance of `FlutterEngine` before looking up the
                 // callback. If we don't, the callback cache won't be initialized and the
                 // lookup will fail.
-                flutterEngine = new FlutterEngine(getApplicationContext());
-
-                FlutterCallbackInformation callbackInfo = FlutterCallbackInformation.lookupCallbackInformation(callbackHandle);
-                if (callbackInfo == null) {
-                    Log.e(TAG, "Fatal: failed to find callback");
+                FlutterCallbackInformation flutterCallback =
+                        FlutterCallbackInformation.lookupCallbackInformation(callbackHandle);
+                if (flutterCallback == null) {
+                    log("Fatal: failed to find callback");
                     return;
                 }
 
                 final String appBundlePath = FlutterInjector.instance().flutterLoader().findAppBundlePath();
                 final AssetManager assets = getApplicationContext().getAssets();
-                flutterEngine.getDartExecutor().executeDartCallback(new DartExecutor.DartCallback(assets, appBundlePath, callbackInfo));
+                backgroundFlutterEngine.getDartExecutor().executeDartCallback(new DartExecutor.DartCallback(assets, appBundlePath, flutterCallback));
             }
+            backgroundChannel = new MethodChannel(backgroundFlutterEngine.getDartExecutor(), "vn.hunghd/downloader_background");
+            backgroundChannel.setMethodCallHandler(this);
         }
-
-        backgroundChannel = new MethodChannel(flutterEngine.getDartExecutor(), "vn.hunghd/downloader_background");
-        backgroundChannel.setMethodCallHandler(this);
     }
 
     @Override
@@ -140,6 +149,22 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
             }
         } else {
             result.notImplemented();
+        }
+    }
+
+    @Override
+    public void onStopped() {
+        Context context = getApplicationContext();
+        dbHelper = TaskDbHelper.getInstance(context);
+        taskDao = new TaskDao(dbHelper);
+
+        String url = getInputData().getString(ARG_URL);
+        String filename = getInputData().getString(ARG_FILE_NAME);
+
+        DownloadTask task = taskDao.loadTask(getId().toString());
+        if (task.status == DownloadStatus.ENQUEUED) {
+            updateNotification(context, filename == null ? url : filename, DownloadStatus.CANCELED, -1, null, true);
+            taskDao.updateTask(getId().toString(), DownloadStatus.CANCELED, lastProgress);
         }
     }
 
@@ -165,12 +190,19 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         msgPaused = res.getString(R.string.flutter_downloader_notification_paused);
         msgComplete = res.getString(R.string.flutter_downloader_notification_complete);
 
-        log("DownloadWorker{url=" + url + ",filename=" + filename + ",savedDir=" + savedDir + ",header=" + headers + ",isResume=" + isResume);
+        DownloadTask task = taskDao.loadTask(getId().toString());
+
+        log("DownloadWorker{url=" + url + ",filename=" + filename + ",savedDir=" + savedDir + ",header=" + headers + ",isResume=" + isResume + ",status=" + (task != null ? task.status : "GONE"));
+
+        // Task has been deleted or cancelled
+        if (task == null || task.status == DownloadStatus.CANCELED) {
+            return Result.success();
+        }
 
         showNotification = getInputData().getBoolean(ARG_SHOW_NOTIFICATION, false);
         clickToOpenDownloadedFile = getInputData().getBoolean(ARG_OPEN_FILE_FROM_NOTIFICATION, false);
+        saveInPublicStorage = getInputData().getBoolean(ARG_SAVE_IN_PUBLIC_STORAGE, false);
 
-        DownloadTask task = taskDao.loadTask(getId().toString());
         primaryId = task.primaryId;
 
         setupNotification(context);
@@ -226,6 +258,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         conn.setRequestProperty("Accept-Encoding", "identity");
         conn.setRequestProperty("Range", "bytes=" + downloadedBytes + "-");
         conn.setDoInput(true);
+
         return downloadedBytes;
     }
 
@@ -235,8 +268,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         Map<String, Integer> visited;
         HttpURLConnection httpConn = null;
         InputStream inputStream = null;
-        FileOutputStream outputStream = null;
-        String saveFilePath;
+        OutputStream outputStream = null;
         String location;
         long downloadedBytes = 0;
         int responseCode;
@@ -278,6 +310,8 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                     case HttpURLConnection.HTTP_MOVED_PERM:
                     case HttpURLConnection.HTTP_SEE_OTHER:
                     case HttpURLConnection.HTTP_MOVED_TEMP:
+                    case 307: /* HTTP_TEMP_REDIRECT */
+                    case 308: /* HTTP_PERM_REDIRECT */
                         log("Response with redirection code");
                         location = httpConn.getHeaderField("Location");
                         log("Location = " + location);
@@ -292,9 +326,9 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
             }
 
             httpConn.connect();
-
+            String contentType;
             if ((responseCode == HttpURLConnection.HTTP_OK || (isResume && responseCode == HttpURLConnection.HTTP_PARTIAL)) && !isStopped()) {
-                String contentType = httpConn.getContentType();
+                contentType = httpConn.getContentType();
                 long contentLength = android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N ? httpConn.getContentLengthLong() : httpConn.getContentLength();
                 log("Content-Type = " + contentType);
                 log("Content-Length = " + contentLength);
@@ -320,7 +354,6 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                         }
                     }
                 }
-                saveFilePath = savedDir + File.separator + filename;
 
                 log("fileName = " + filename);
 
@@ -329,12 +362,35 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                 // opens input stream from the HTTP connection
                 inputStream = httpConn.getInputStream();
 
+
+                String savedFilePath;
                 // opens an output stream to save into file
-                outputStream = new FileOutputStream(saveFilePath, isResume);
+                // there are two case:
+                if (isResume) {
+                    // 1. continue downloading (append data to partial downloaded file)
+                    savedFilePath = savedDir + File.separator + filename;
+                    outputStream = new FileOutputStream(savedFilePath, true);
+                } else {
+                    // 2. new download, create new file
+                    // there are two case according to Android SDK version and save path
+                    // From Android 11 onwards, file is only downloaded to app-specific directory (internal storage)
+                    // or public shared download directory (external storage).
+                    // The second option will ignore `savedDir` parameter.
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q && saveInPublicStorage) {
+                        Uri uri = createFileInPublicDownloadsDir(filename, contentType);
+                        savedFilePath = getMediaStoreEntryPathApi29(uri);
+                        outputStream = context.getContentResolver().openOutputStream(uri, "w");
+                    } else {
+                        File file = createFileInAppSpecificDir(filename, savedDir);
+                        savedFilePath = file.getPath();
+                        outputStream = new FileOutputStream(file, false);
+                    }
+                }
 
                 long count = downloadedBytes;
                 int bytesRead = -1;
                 byte[] buffer = new byte[BUFFER_SIZE];
+                // using isStopped() to monitor canceling task
                 while ((bytesRead = inputStream.read(buffer)) != -1 && !isStopped()) {
                     count += bytesRead;
                     int progress = (int) ((count * 100) / (contentLength + downloadedBytes));
@@ -343,13 +399,14 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                     if ((lastProgress == 0 || progress > lastProgress + STEP_UPDATE || progress == 100)
                             && progress != lastProgress) {
                         lastProgress = progress;
-                        updateNotification(context, filename, DownloadStatus.RUNNING, progress, null, false);
 
                         // This line possibly causes system overloaded because of accessing to DB too many ?!!!
                         // but commenting this line causes tasks loaded from DB missing current downloading progress,
                         // however, this missing data should be temporary and it will be updated as soon as
                         // a new bunch of data fetched and a notification sent
                         taskDao.updateTask(getId().toString(), DownloadStatus.RUNNING, progress);
+
+                        updateNotification(context, filename, DownloadStatus.RUNNING, progress, null, false);
                     }
                 }
 
@@ -359,37 +416,43 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                 int storage = ContextCompat.checkSelfPermission(getApplicationContext(), android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
                 PendingIntent pendingIntent = null;
                 if (status == DownloadStatus.COMPLETE) {
-                    if (isImageOrVideoFile(contentType) && isExternalStoragePath(saveFilePath)) {
-                        addImageOrVideoToGallery(filename, saveFilePath, getContentTypeWithoutCharset(contentType));
+                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+                        if (isImageOrVideoFile(contentType) && isExternalStoragePath(savedFilePath)) {
+                            addImageOrVideoToGallery(filename, savedFilePath, getContentTypeWithoutCharset(contentType));
+                        }
                     }
 
-                    if (clickToOpenDownloadedFile && storage == PackageManager.PERMISSION_GRANTED) {
-                        Intent intent = IntentUtils.validatedFileIntent(getApplicationContext(), saveFilePath, contentType);
+                    if (clickToOpenDownloadedFile) {
+                        if (android.os.Build.VERSION.SDK_INT < Build.VERSION_CODES.Q && storage != PackageManager.PERMISSION_GRANTED)
+                            return;
+                        Intent intent = IntentUtils.validatedFileIntent(getApplicationContext(), savedFilePath, contentType);
                         if (intent != null) {
-                            log("Setting an intent to open the file " + saveFilePath);
-                            pendingIntent = PendingIntent.getActivity(getApplicationContext(), 0, intent, PendingIntent.FLAG_CANCEL_CURRENT);
+                            log("Setting an intent to open the file " + savedFilePath);
+                            int flags = Build.VERSION.SDK_INT >= Build.VERSION_CODES.S ? PendingIntent.FLAG_CANCEL_CURRENT | PendingIntent.FLAG_IMMUTABLE : PendingIntent.FLAG_CANCEL_CURRENT;
+                            pendingIntent = PendingIntent.getActivity(getApplicationContext(), 0, intent, flags);
                         } else {
-                            log("There's no application that can open the file " + saveFilePath);
+                            log("There's no application that can open the file " + savedFilePath);
                         }
                     }
                 }
-                updateNotification(context, filename, status, progress, pendingIntent, true);
                 taskDao.updateTask(getId().toString(), status, progress);
+                updateNotification(context, filename, status, progress, pendingIntent, true);
 
                 log(isStopped() ? "Download canceled" : "File downloaded");
             } else {
                 DownloadTask task = taskDao.loadTask(getId().toString());
                 int status = isStopped() ? (task.resumable ? DownloadStatus.PAUSED : DownloadStatus.CANCELED) : DownloadStatus.FAILED;
-                updateNotification(context, filename == null ? fileURL : filename, status, -1, null, true);
                 taskDao.updateTask(getId().toString(), status, lastProgress);
+                updateNotification(context, filename == null ? fileURL : filename, status, -1, null, true);
                 log(isStopped() ? "Download canceled" : "Server replied HTTP code: " + responseCode);
             }
         } catch (IOException e) {
-            updateNotification(context, filename == null ? fileURL : filename, DownloadStatus.FAILED, -1, null, true);
             taskDao.updateTask(getId().toString(), DownloadStatus.FAILED, lastProgress);
+            updateNotification(context, filename == null ? fileURL : filename, DownloadStatus.FAILED, -1, null, true);
             e.printStackTrace();
         } finally {
             if (outputStream != null) {
+                outputStream.flush();
                 try {
                     outputStream.close();
                 } catch (IOException e) {
@@ -407,6 +470,76 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                 httpConn.disconnect();
             }
         }
+    }
+
+    /**
+     * Create a file using java.io API
+     */
+    private File createFileInAppSpecificDir(String filename, String savedDir) {
+        File newFile = new File(savedDir, filename);
+        try {
+            boolean rs = newFile.createNewFile();
+            if (rs) {
+                return newFile;
+            } else {
+                logError("It looks like you are trying to save file in public storage but not setting 'saveInPublicStorage' to 'true'");
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+            logError("Create a file using java.io API failed ");
+        }
+        return null;
+    }
+
+    /**
+     * Create a file inside the Download folder using MediaStore API
+     */
+    @RequiresApi(Build.VERSION_CODES.Q)
+    private Uri createFileInPublicDownloadsDir(String filename, String mimeType) {
+        Uri collection = MediaStore.Downloads.EXTERNAL_CONTENT_URI;
+        ContentValues values = new ContentValues();
+        values.put(MediaStore.Downloads.DISPLAY_NAME, filename);
+        values.put(MediaStore.Downloads.MIME_TYPE, mimeType);
+        values.put(MediaStore.Downloads.RELATIVE_PATH, Environment.DIRECTORY_DOWNLOADS);
+        ContentResolver contentResolver = getApplicationContext().getContentResolver();
+        try {
+            return contentResolver.insert(collection, values);
+        } catch (Exception e) {
+            e.printStackTrace();
+            logError("Create a file using MediaStore API failed.");
+        }
+        return null;
+    }
+
+    /**
+     * Get a path for a MediaStore entry as it's needed when calling MediaScanner
+     */
+    private String getMediaStoreEntryPathApi29(Uri uri) {
+        try (Cursor cursor = getApplicationContext().getContentResolver().query(
+                uri,
+                new String[]{MediaStore.Files.FileColumns.DATA},
+                null,
+                null,
+                null
+        )) {
+            if (cursor == null)
+                return null;
+            if (!cursor.moveToFirst())
+                return null;
+            return cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Files.FileColumns.DATA));
+        } catch (IllegalArgumentException e) {
+            e.printStackTrace();
+            logError("Get a path for a MediaStore failed");
+            return null;
+        }
+    }
+
+    void scanFilePath(String path, String mimeType, CallbackUri callback) {
+        MediaScannerConnection.scanFile(
+                getApplicationContext(),
+                new String[]{path},
+                new String[]{mimeType},
+                (path1, uri) -> callback.invoke(uri));
     }
 
     private void cleanUp() {
@@ -649,5 +782,15 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         if (debug) {
             Log.d(TAG, message);
         }
+    }
+
+    private void logError(String message) {
+        if (debug) {
+            Log.e(TAG, message);
+        }
+    }
+
+    public interface CallbackUri {
+        void invoke(Uri uri);
     }
 }